app/matching/requirements.txt
fastapi
numpy
pydantic
pytest
sentence-transformers
sqlalchemy
uvicorn
httpx
python-dotenv
google-cloud-bigquery
